MEDIUM/WIP: ssl: process the 'crt' path in a pre_check function
Stop calling ssl_sock_load_cert() from bind_parse_crt() in order to
separate the generation of SSL_CTX from the parsing.

This is supposed to remove part of the dependency to OpenSSL in the
cfgparse-ssl.c file.

Probably takes longer than before with a huge list of proxies, this is
fine for now but must be moved later where the proxies are already
iterated.

There is some visible change upon error, the error or warning emitted
don't know the current section so it displays the filename and line
only and not the section anymore which is probably enough.